refactor: moved auth endpoints to auth module
