+++ api/resource/user/user.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/go-playground/validator/v10"
-	"github.com/golang-jwt/jwt"
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
 	"github.com/ushiradineth/cron-be/models"
@@ -160,48 +159,6 @@ func (api *API) Put(w http.ResponseWriter, r *http.Request) {
 	util.HTTPResponse(w, user)
 }
 
-// @Summary		Update User Password
-// @Description	Update authenticated User Password with the parameters sent with the request based on the JWT
-// @Tags			User
-// @Accept			json
-// @Produce		json
-// @Param			Query	query		PutPasswordQueryParams	true	"PutPasswordQueryParams"
-// @Success		200		{object}	util.Response{data=string}
-// @Failure		400		{object}	util.Error
-// @Failure		401		{object}	util.Error
-// @Failure		500		{object}	util.Error
-// @Security		BearerAuth
-// @Router			/user/auth/password [put]
-func (api *API) PutPassword(w http.ResponseWriter, r *http.Request) {
-	query := PutPasswordQueryParams{
-		Password: r.FormValue("password"),
-	}
-
-	if err := api.validator.Struct(query); err != nil {
-		util.GenericValidationError(w, err)
-		return
-	}
-
-	user := util.GetUserFromJWT(r, w, api.db)
-	if user == nil {
-		return
-	}
-
-	password, err := util.HashPassword(query.Password)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	_, err = api.db.Exec("UPDATE users SET password=$1 WHERE id=$2", password, user.ID)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	util.HTTPResponse(w, "Password has being updated")
-}
-
 // @Summary		Delete User
 // @Description	Delete authenticated User based on the JWT
 // @Tags			User
@@ -237,124 +194,4 @@ func (api *API) Delete(w http.ResponseWriter, r *http.Request) {
 	}
 
 	util.HTTPResponse(w, "User has been successfully deleted")
-}
-
-// @Summary		Authenticate User
-// @Description	Authenticated User with the parameters sent with the request
-// @Tags			User
-// @Accept			json
-// @Produce		json
-// @Param			Query	query		AuthenticateQueryParams	true	"AuthenticateQueryParams"
-// @Success		200		{object}	util.Response{data=AuthenticateResponse}
-// @Failure		400		{object}	util.Error
-// @Failure		401		{object}	util.Error
-// @Failure		500		{object}	util.Error
-// @Router			/user/auth [post]
-func (api *API) Authenticate(w http.ResponseWriter, r *http.Request) {
-	query := AuthenticateQueryParams{
-		Email:    r.FormValue("email"),
-		Password: r.FormValue("password"),
-	}
-
-	if err := api.validator.Struct(query); err != nil {
-		util.GenericValidationError(w, err)
-		return
-	}
-
-	user := util.GetUser(w, query.Email, api.db)
-	if user == nil {
-		return
-	}
-
-	valid := util.CheckPasswordHash(query.Password, user.Password)
-
-	if !valid {
-		util.HTTPError(w, http.StatusUnauthorized, "Invalid Credentials", util.StatusFail)
-		return
-	}
-
-	accessToken, err := util.NewAccessToken(user.ID, user.Name, user.Email)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	refreshTokenClaim := jwt.StandardClaims{
-		IssuedAt:  time.Now().Unix(),
-		ExpiresAt: time.Now().Add(time.Hour * 48).Unix(),
-	}
-
-	refreshToken, err := util.NewRefreshToken(refreshTokenClaim)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	user.Password = "redacted"
-
-	response := AuthenticateResponse{
-		User:         *user,
-		AccessToken:  accessToken,
-		RefreshToken: refreshToken,
-	}
-
-	util.HTTPResponse(w, response)
-}
-
-// @Summary		Refresh Access Token
-// @Description	Refresh Access Token User with the parameters sent with the request based on the request based on the JWT
-// @Tags			User
-// @Accept			json
-// @Produce		json
-// @Param			Query	query		RefreshTokenQueryParams	true	"RefreshTokenQueryParams"
-// @Success		200		{object}	util.Response{data=RefreshTokenResponse}
-// @Failure		400		{object}	util.Error
-// @Failure		401		{object}	util.Error
-// @Failure		500		{object}	util.Error
-// @Security		BearerAuth
-// @Router			/user/auth/refresh [post]
-func (api *API) RefreshToken(w http.ResponseWriter, r *http.Request) {
-	accessToken, err := util.GetJWT(r)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	query := RefreshTokenQueryParams{
-		RefreshToken: r.FormValue("refresh_token"),
-	}
-
-	if err := api.validator.Struct(query); err != nil {
-		util.GenericValidationError(w, err)
-		return
-	}
-
-	accessTokenClaim, err := util.ParseExpiredAccessToken(accessToken)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	user := util.GetUser(w, accessTokenClaim.Email, api.db)
-	if user == nil {
-		return
-	}
-
-	_, errr := util.ParseRefreshToken(query.RefreshToken)
-	if errr != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	newAccessToken, err := util.NewAccessToken(user.ID, user.Name, user.Email)
-	if err != nil {
-		util.GenericServerError(w, err)
-		return
-	}
-
-	response := RefreshTokenResponse{
-		AccessToken: newAccessToken,
-	}
-
-	util.HTTPResponse(w, response)
 }
