+++ api/router/router.go
@@ -6,6 +6,7 @@ import (
 	"github.com/go-playground/validator/v10"
 	"github.com/jmoiron/sqlx"
 	"github.com/swaggo/http-swagger/v2"
+	"github.com/ushiradineth/cron-be/api/resource/auth"
 	"github.com/ushiradineth/cron-be/api/resource/event"
 	"github.com/ushiradineth/cron-be/api/resource/health"
 	"github.com/ushiradineth/cron-be/api/resource/user"
@@ -21,9 +22,11 @@ func New(db *sqlx.DB, v *validator.Validate) http.Handler {
 	mux.HandleFunc("POST /user", userAPI.Post)
 	mux.HandleFunc("PUT /user", userAPI.Put)
 	mux.HandleFunc("DELETE /user", userAPI.Delete)
-	mux.HandleFunc("POST /user/auth", userAPI.Authenticate)
-	mux.HandleFunc("POST /user/auth/refresh", userAPI.RefreshToken)
-	mux.HandleFunc("PUT /user/auth/password", userAPI.PutPassword)
+
+	authAPI := auth.New(db, v)
+	mux.HandleFunc("POST /auth/login", authAPI.Authenticate)
+	mux.HandleFunc("POST /auth/refresh", authAPI.RefreshToken)
+	mux.HandleFunc("PUT /auth/reset-password", authAPI.PutPassword)
 
 	eventAPI := event.New(db, v)
 	mux.HandleFunc("GET /event/{event_id}", eventAPI.Get)
